[fix] renamed download/upload and init commands to ninox database download , ninox database upload and ninox project init ; commands
